Removed ComponentFactory and renamed ComponentAdder to ComponentFactory
Fixed a long standing typo JSONPasrser is now JSONParser.

Most of the functionality of ComponentFactory was moved into JSONParser.

ComponentAdder is in fact a factory and the name is being changed to match its function.

Cleaned up an include (first of many) in an effort to reduce compile times.
component.hpp no longer directly includes physics.hpp.

Updated JSON format for entities.